Truncate filenames and filepaths more intelligently
Instead of asking the user to supply a different name for a file
whose name is too long, the program truncates the file to its
maximum length. Note that this may not work well because the model
params may not be set well and because some parts of the code have
not been set to truncate the names of the files they try to find.
Both of these problems will be fixed in coming commits.